requirements.txt
torch
torchvision
facenet-pytorch
opencv-python
numpy
Pillow
pandas
matplotlib
seaborn
scikit-learn
tqdm
jupyter
fastapi
uvicorn
python-multipart
timm
albumentations
streamlit
requests